refactor: add a function module of torch
